+++ contracts/mock/base/escrow/dapps/UniswapV2Router02Mock.sol
@@ -3,12 +3,9 @@ pragma solidity 0.5.17;
 import "../../../../providers/uniswap/IUniswapV2Router02.sol";
 import "../../../../mock/token/ERC20Mock.sol";
 
-
 contract UniswapV2Router02Mock is IUniswapV2Router02 {
-
     uint24 public constant DONT_ALTER_BALANCE = 999999;
     uint24 public constant SIMULATE_UNISWAP_RESPONSE_ERROR = 777777;
-
 
     /**
         @notice This function swaps tokens for tokens.
@@ -20,16 +17,9 @@ contract UniswapV2Router02Mock is IUniswapV2Router02 {
         address[] calldata path,
         address to,
         uint256 deadline
-    ) external returns (uint256[] memory amounts) 
-    {
-        return _swap(
-           amountOutMin,
-           path,
-           to,
-           deadline
-       );
+    ) external returns (uint256[] memory amounts) {
+        return _swap(amountOutMin, path, to, deadline);
     }
-
 
     /**
         @notice This function swaps tokens for WETH tokens.
@@ -42,14 +32,8 @@ contract UniswapV2Router02Mock is IUniswapV2Router02 {
         address[] calldata path,
         address to,
         uint256 deadline
-    ) external returns (uint256[] memory amounts)
-    {
-        return _swap(
-           amountOutMin,
-           path,
-           to,
-           deadline
-       );
+    ) external returns (uint256[] memory amounts) {
+        return _swap(amountOutMin, path, to, deadline);
     }
 
     /**
@@ -61,37 +45,27 @@ contract UniswapV2Router02Mock is IUniswapV2Router02 {
         address[] calldata path,
         address to,
         uint256 deadline
-    ) external payable returns (uint256[] memory amounts)
-    {
-       return _swap(
-           amountOutMin,
-           path,
-           to,
-           deadline
-       );
+    ) external payable returns (uint256[] memory amounts) {
+        return _swap(amountOutMin, path, to, deadline);
     }
 
-    function _swap( 
+    function _swap(
         uint256 amountOutMin,
         address[] memory path,
         address to,
         uint256
-    ) 
-        internal 
-        returns (uint256[] memory amounts)
-    {
+    ) internal returns (uint256[] memory amounts) {
         if (SIMULATE_UNISWAP_RESPONSE_ERROR == amountOutMin) {
             amounts = new uint256[](path.length + 1);
         } else {
             amounts = new uint256[](path.length);
         }
-        for (uint i; i < path.length; i++) {
+        for (uint256 i; i < path.length; i++) {
             amounts[i] = amountOutMin;
         }
-        if (!(DONT_ALTER_BALANCE == amountOutMin)){
-            ERC20Mock destinationToken = ERC20Mock(path[ path.length - 1 ]);
+        if (!(DONT_ALTER_BALANCE == amountOutMin)) {
+            ERC20Mock destinationToken = ERC20Mock(path[path.length - 1]);
             destinationToken.mint(to, amountOutMin);
         }
     }
-
 }
